move storage to src
