+++ dxcherry/src/extension.ts
@@ -1,7 +1,7 @@
 import { commands, ExtensionContext, window } from 'vscode';
 import TreeDataProvider from './tree/dataProvider';
 import { Action } from './actions';
-import Storage from './reviewers/storage';
+import Storage from './storage';
 import GithubClient from './github/client';
 import Git from './git/client';
